Progress made on integrating libsndfile - at least stuff compiles. Also fixed bug in missionhandler with xml file extension
